Use spans of std::byte in serialize
Summary:
This switches .read() and .write() to take spans of bytes.

This is a backport of [[https://github.com/bitcoin/bitcoin/pull/23438 | core#23438]]
https://github.com/bitcoin/bitcoin/pull/23438/commits/fa24493d6394b3a477535f480664c9596f18e3c5

Depends ib D15934

Test Plan:
`ninja all check-all bitcoin-fuzzers bench-bitcoin`

Run the seeder for a few minutes, make sure it finds nodes

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D15935

+++ src/seeder/test/message_writer_tests.cpp
@@ -23,7 +23,7 @@ static void CheckMessage(CDataStream &expectedMessage, std::string command,
     MessageWriter::WriteMessage(message, command, std::forward<Args>(args)...);
     BOOST_CHECK_EQUAL(message.size(), expectedMessage.size());
     for (size_t i = 0; i < message.size(); i++) {
-        BOOST_CHECK_EQUAL(message[i], expectedMessage[i]);
+        BOOST_CHECK_EQUAL(uint8_t(message[i]), uint8_t(expectedMessage[i]));
     }
 }
